test/scripts/requirements_simple.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
PyMuPDF==1.23.8
Pillow==10.1.0
python-multipart==0.0.6
sentence-transformers==2.7.0
transformers==4.44.0
huggingface-hub==0.24.0
torch>=2.2.0
numpy==1.24.3
scikit-learn==1.3.2
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-json-logger==2.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0